+++ RELEASENOTES.md
@@ -6,6 +6,9 @@
 *   ExoPlayer:
 *   Transformer:
 *   Track Selection:
+    *   Add `DefaultTrackSelector.Parameters.allowAudioNonSeamlessAdaptiveness`
+        to explicitly allow or disallow non-seamless adaptation. The default
+        stays at its current behavior of `true`.
 *   Extractors:
 *   Audio:
 *   Video: